% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reco_gauss.R
\name{csmvn}
\alias{csmvn}
\title{Cross-sectional Gaussian probabilistic reconciliation}
\usage{
csmvn(
  base,
  agg_mat,
  cons_mat,
  comb = "ols",
  comb_base = comb,
  res = NULL,
  approach = "proj",
  reduce_form = FALSE,
  ...
)
}
\arguments{
\item{base}{A (\eqn{h \times n}) numeric matrix or multivariate time series
(\code{mts} class) containing the base forecasts to be reconciled; \eqn{h}
is the forecast horizon, and \eqn{n} is the total number of time series
(\eqn{n = n_a + n_b}).}

\item{agg_mat}{A (\eqn{n_a \times n_b}) numeric matrix representing the
cross-sectional aggregation matrix. It maps the \eqn{n_b} bottom-level
(free) variables into the \eqn{n_a} upper (constrained) variables.}

\item{cons_mat}{A (\eqn{n_a \times n}) numeric matrix representing the
cross-sectional zero constraints: each row represents a constraint
equation, and each column represents a variable. The matrix can be of full
rank, meaning the rows are linearly independent, but this is not a strict
requirement, as the function allows for redundancy in the constraints.}

\item{comb}{A string specifying the reconciliation method. For a complete
list, see \link{cscov}.}

\item{comb_base}{A string specifying the reconciliation method. For a complete list, see \link{cscov}.}

\item{res}{An (\eqn{N \times n}) optional numeric matrix containing the
in-sample residuals. This matrix is used to compute some covariance
matrices.}

\item{approach}{A string specifying the approach used to compute the reconciled
mean and covariance matrix. Options include:
\itemize{
\item "\code{proj}" (\emph{default}): Projection approach according to Byron (1978, 1979).
\item "\code{strc}": Structural approach as proposed by Hyndman et al. (2011).
}}

\item{reduce_form}{A logical parameter indicating whether the function should return the
full distribution (\code{FALSE}, \emph{default}) or only the distribution corresponding
to the bottom-level time series (\code{TRUE}).}

\item{...}{
  Arguments passed on to \code{\link[=cscov]{cscov}}
  \describe{
    \item{\code{mse}}{If \code{TRUE} (\emph{default}) the residuals used to compute the
covariance matrix are not mean-corrected.}
    \item{\code{shrink_fun}}{Shrinkage function of the covariance matrix,
\link{shrink_estim} (\emph{default}).}
  }}
}
\value{
A \link[distributional:dist_multivariate_normal]{distributional::dist_multivariate_normal} object.
}
\description{
Cross-sectional Gaussian probabilistic reconciliation
}
\examples{
set.seed(123)
# (2 x 3) base forecasts matrix (simulated), Z = X + Y
base <- matrix(rnorm(6, mean = c(20, 10, 10)), 2, byrow = TRUE)
# (10 x 3) in-sample residuals matrix (simulated)
res <- t(matrix(rnorm(n = 30), nrow = 3))

# Aggregation matrix for Z = X + Y
A <- t(c(1,1))
reco_dist <- csmvn(base = base, agg_mat = A, comb = "shr", res = res)

}
\references{
Byron, R.P. (1978), The estimation of large social account matrices,
\emph{Journal of the Royal Statistical Society, Series A}, 141, 3, 359-367.
\doi{10.2307/2344807}

Byron, R.P. (1979), Corrigenda: The estimation of large social account matrices,
\emph{Journal of the Royal Statistical Society, Series A}, 142(3), 405.
\doi{10.2307/2982515}

Girolimetto, D., Athanasopoulos, G., Di Fonzo, T. and Hyndman, R.J. (2024),
Cross-temporal probabilistic forecast reconciliation: Methodological and
practical issues. \emph{International Journal of Forecasting}, 40, 3, 1134-1151.
\doi{10.1016/j.ijforecast.2023.10.003}

Hyndman, R.J., Ahmed, R.A., Athanasopoulos, G. and Shang, H.L. (2011),
Optimal combination forecasts for hierarchical time series,
\emph{Computational Statistics & Data Analysis}, 55, 9, 2579-2589.
\doi{10.1016/j.csda.2011.03.006}

Panagiotelis, A., Gamakumara, P., Athanasopoulos, G. and Hyndman, R.J. (2023),
Probabilistic forecast reconciliation: Properties, evaluation and score optimisation,
\emph{European Journal of Operational Research} 306(2), 693–706.
\doi{http://dx.doi.org/10.1016/j.ejor.2022.07.040}
}
\seealso{
Probabilistic reconciliation:
\code{\link{cssmp}()},
\code{\link{ctmvn}()},
\code{\link{ctsmp}()},
\code{\link{temvn}()},
\code{\link{tesmp}()}

Cross-sectional framework:
\code{\link{csboot}()},
\code{\link{csbu}()},
\code{\link{cscov}()},
\code{\link{cslcc}()},
\code{\link{csmo}()},
\code{\link{csrec}()},
\code{\link{cssmp}()},
\code{\link{cstd}()},
\code{\link{cstools}()}
}
\concept{Framework: cross-sectional}
\concept{Reco: probabilistic}
