% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reco_gauss.R
\name{ctmvn}
\alias{ctmvn}
\title{Cross-temporal Gaussian probabilistic reconciliation}
\usage{
ctmvn(base, agg_order, agg_mat, cons_mat, comb = "ols", comb_base = comb,
      res = NULL, tew = "sum", approach = "proj", reduce_form = FALSE, ...)
}
\arguments{
\item{base}{A (\eqn{n \times h(k^\ast+m)}) numeric matrix containing the base
forecasts to be reconciled; \eqn{n} is the total number of variables,
\eqn{m} is the maximum aggregation order, and \eqn{k^\ast} is the sum of a
chosen subset of the \eqn{p - 1} factors of \eqn{m} (excluding \eqn{m}
itself), and \eqn{h} is the forecast horizon for the lowest frequency time
series. The row identifies a time series, and the forecasts in each row are
ordered from the lowest frequency (most temporally aggregated) to the
highest frequency.}

\item{agg_mat}{A (\eqn{n_a \times n_b}) numeric matrix representing the
cross-sectional aggregation matrix. It maps the \eqn{n_b} bottom-level
(free) variables into the \eqn{n_a} upper (constrained) variables.}

\item{cons_mat}{A (\eqn{n_a \times n}) numeric matrix representing the
cross-sectional zero constraints: each row represents a constraint
equation, and each column represents a variable. The matrix can be of full
rank, meaning the rows are linearly independent, but this is not a strict
requirement, as the function allows for redundancy in the constraints.}

\item{agg_order}{Highest available sampling frequency per seasonal cycle
(max. order of temporal aggregation, \eqn{m}), or a vector representing a
subset of \eqn{p} factors of \eqn{m}.}

\item{comb}{A string specifying the reconciliation method. For a complete
list, see \link{ctcov}.}

\item{comb_base}{A string specifying the reconciliation method. For a complete list, see \link{ctcov}.}

\item{res}{A (\eqn{n \times N(k^\ast+m)}) optional numeric matrix containing
the in-sample residuals ordered from the lowest frequency to the highest
frequency (columns) for each variable (rows). This matrix is used to
compute some covariance matrices.}

\item{tew}{A string specifying the type of temporal aggregation. Options
include: "\code{sum}" (simple summation, \emph{default}), "\code{avg}"
(average), "\code{first}" (first value of the period), and "\code{last}"
(last value of the period).}

\item{approach}{A string specifying the approach used to compute the
reconciled forecasts. Options include:
\itemize{
\item "\code{proj}" (\emph{default}): Projection approach according
to Byron (1978, 1979).
\item "\code{strc}": Structural approach as proposed by Hyndman
et al. (2011).
\item "\code{proj_osqp}": Numerical solution using
\href{https://osqp.org/}{\pkg{osqp}} for projection approach.
\item "\code{strc_osqp}": Numerical solution using
\href{https://osqp.org/}{\pkg{osqp}} for structural approach.
}}

\item{reduce_form}{A logical parameter indicating whether the function should return the
full distribution (\code{FALSE}, \emph{default}) or only the distribution corresponding
to the high-frequency bottom time series (\code{TRUE}).}

\item{...}{
  Arguments passed on to \code{\link[=ctcov]{ctcov}}
  \describe{
    \item{\code{mse}}{If \code{TRUE} (\emph{default}) the residuals used to compute the
covariance matrix are not mean-corrected.}
    \item{\code{shrink_fun}}{Shrinkage function of the covariance matrix,
\link{shrink_estim} (\emph{default}).}
  }}
}
\value{
A \link[distributional:dist_multivariate_normal]{distributional::dist_multivariate_normal} object.
}
\description{
This function performs cross-temporal probabilistic forecast reconciliation
assuming a multivariate normal base forecast distribution (Girolimetto et
al., 2024) for linearly constrained multiple time series observed across both
cross-sectional and temporal dimensions (Di Fonzo and Girolimetto, 2023).
}
\examples{
set.seed(123)
# (3 x 7) base forecasts matrix (simulated), Z = X + Y and m = 4
base <- rbind(rnorm(7, rep(c(20, 10, 5), c(1, 2, 4))),
              rnorm(7, rep(c(10, 5, 2.5), c(1, 2, 4))),
              rnorm(7, rep(c(10, 5, 2.5), c(1, 2, 4))))
# (3 x 70) in-sample residuals matrix (simulated)
res <- rbind(rnorm(70), rnorm(70), rnorm(70))
A <- t(c(1,1))
reco_dist <- ctmvn(base = base, res = res, agg_mat = A, agg_order = 4)

}
\references{
Byron, R.P. (1978), The estimation of large social account matrices,
\emph{Journal of the Royal Statistical Society, Series A}, 141, 3, 359-367.
\doi{10.2307/2344807}

Byron, R.P. (1979), Corrigenda: The estimation of large social account matrices,
\emph{Journal of the Royal Statistical Society, Series A}, 142(3), 405.
\doi{10.2307/2982515}

Girolimetto, D., Athanasopoulos, G., Di Fonzo, T. and Hyndman, R.J. (2024),
Cross-temporal probabilistic forecast reconciliation: Methodological and
practical issues. \emph{International Journal of Forecasting}, 40, 3, 1134-1151.
\doi{10.1016/j.ijforecast.2023.10.003}

Hyndman, R.J., Ahmed, R.A., Athanasopoulos, G. and Shang, H.L. (2011),
Optimal combination forecasts for hierarchical time series,
\emph{Computational Statistics & Data Analysis}, 55, 9, 2579-2589.
\doi{10.1016/j.csda.2011.03.006}

Panagiotelis, A., Gamakumara, P., Athanasopoulos, G. and Hyndman, R.J. (2023),
Probabilistic forecast reconciliation: Properties, evaluation and score optimisation,
\emph{European Journal of Operational Research} 306(2), 693–706.
\doi{http://dx.doi.org/10.1016/j.ejor.2022.07.040}
}
\seealso{
Probabilistic reconciliation:
\code{\link{csmvn}()},
\code{\link{cssmp}()},
\code{\link{ctsmp}()},
\code{\link{temvn}()},
\code{\link{tesmp}()}

Cross-temporal framework:
\code{\link{ctboot}()},
\code{\link{ctbu}()},
\code{\link{ctcov}()},
\code{\link{ctlcc}()},
\code{\link{ctmo}()},
\code{\link{ctrec}()},
\code{\link{ctsmp}()},
\code{\link{cttd}()},
\code{\link{cttools}()},
\code{\link{iterec}()},
\code{\link{tcsrec}()}
}
\concept{Framework: cross-temporal}
\concept{Reco: probabilistic}
