% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htsrec.R
\name{htsrec}
\alias{htsrec}
\title{Cross-sectional (contemporaneous) forecast reconciliation}
\usage{
htsrec(basef, comb, C, res, Ut, nb, mse = TRUE, corpcor = FALSE, W,
       type = "M", sol = "direct", nn = FALSE, keep = "list",
       settings = osqpSettings(verbose = FALSE, eps_abs = 1e-5,
       eps_rel = 1e-5, polish_refine_iter = 100, polish=TRUE))
}
\arguments{
\item{basef}{(\code{h x n}) matrix of base forecasts to be reconciled;
\code{h} is the forecast horizon and \code{n} is the total number of time series.}

\item{comb}{Type of the reconciliation. Except for Bottom-up, each
option corresponds to a specified (\code{n x n}) covariance matrix:
\itemize{
    \item \bold{bu} (Bottom-up);
    \item \bold{ols} (Identity);
    \item \bold{struc} (Structural variances);
    \item \bold{wls} (Series variances) - uses res;
    \item \bold{shr} (Shrunk covariance matrix - MinT-shr) - uses res;
    \item \bold{sam} (Sample covariance matrix - MinT-sam) - uses res;
    \item \bold{w} use your personal matrix W in param \code{W}.
  }}

\item{C}{(\code{na x nb}) cross-sectional (contemporaneous) matrix mapping the bottom
level series into the higher level ones.}

\item{res}{(\code{N x n}) in-sample residuals matrix needed when \code{comb =}
\code{\{"wls",} \code{"shr",} \code{"sam"\}}. The columns must be in
the same order as \code{basef}.}

\item{Ut}{Zero constraints cross-sectional (contemporaneous) kernel matrix
\eqn{(\textbf{U}'\textbf{Y} = \mathbf{0})}{} spanning the null space valid for the reconciled
forecasts. It can be used instead of parameter \code{C}, but in this case \code{nb} (n = na + nb) is needed. If
the hierarchy admits a structural representation, \code{Ut} has dimension (\code{na x n}).}

\item{nb}{Number of bottom time series; if \code{C} is present, \code{nb} is not used.}

\item{mse}{Logical value: \code{TRUE} (\emph{default}) calculates the
covariance matrix of the in-sample residuals (when necessary) according to the original
\pkg{hts} and \pkg{thief} formulation: no mean correction, T as denominator.}

\item{corpcor}{Logical value: \code{TRUE} if \pkg{corpcor} (\enc{Sch채fer}{Schafer} et
al., 2017) must be used to shrink the sample covariance matrix according to
\enc{Sch채fer}{Schafer} and Strimmer (2005), otherwise the function uses the same
implementation as package \pkg{hts}.}

\item{W}{This option permits to directly enter the covariance matrix:
\enumerate{
  \item \code{W} must be a p.d. (\code{n x n}) matrix;
  \item if \code{comb} is different from "\code{w}", \code{W} is not used.
}}

\item{type}{Approach used to compute the reconciled forecasts: \code{"M"} for
the projection approach with matrix M (\emph{default}), or \code{"S"} for the
structural approach with summing matrix S.}

\item{sol}{Solution technique for the reconciliation problem: either \code{"direct"} (\emph{default}) for the direct
solution or \code{"osqp"} for the numerical solution (solving a linearly constrained quadratic
program using \code{\link[osqp]{solve_osqp}}).}

\item{nn}{Logical value: \code{TRUE} if non-negative reconciled forecasts are wished.}

\item{keep}{Return a list object of the reconciled forecasts at all levels.}

\item{settings}{Settings for \pkg{osqp} (object \code{\link[osqp]{osqpSettings}}). The default options
are: \code{verbose = FALSE}, \code{eps_abs = 1e-5}, \code{eps_rel = 1e-5},
\code{polish_refine_iter = 100} and \code{polish = TRUE}. For details, see the
\href{https://osqp.org/}{\pkg{osqp} documentation} (Stellato et al., 2019).}
}
\value{
If the parameter \code{keep} is equal to \code{"recf"}, then the function
returns only the (\code{h x n}) reconciled forecasts matrix, otherwise (\code{keep="all"})
it returns a list that mainly depends on what type of representation (\code{type})
and methodology (\code{sol}) have been used:
\item{\code{recf}}{(\code{h x n}) reconciled forecasts matrix.}
\item{\code{W}}{Covariance matrix used for forecast reconciliation.}
\item{\code{nn_check}}{Number of negative values (if zero, there are no values below zero).}
\item{\code{rec_check}}{Logical value: has the hierarchy been respected?}
\item{\code{M} (\code{type="M"} and \code{type="direct"})}{Projection matrix (projection approach)}
\item{\code{G} (\code{type="S"} and \code{type="direct"})}{Projection matrix (structural approach).}
\item{\code{S} (\code{type="S"} and \code{type="direct"})}{Cross-sectional summing matrix, \eqn{\textbf{S}}{S}.}
\item{\code{info} (\code{type="osqp"})}{matrix with information in columns
for each forecast horizon \code{h} (rows): run time (\code{run_time}) number of iteration,
norm of primal residual (\code{pri_res}), status of osqp's solution (\code{status}) and
polish's status (\code{status_polish}).}

Only if \code{comb = "bu"}, the function returns \code{recf}, \code{S} and \code{M}.
}
\description{
Cross-sectional (contemporaneous) forecast reconciliation of hierarchical and
grouped time series. The reconciled forecasts are calculated either through a
projection approach (Byron, 1978, see also van Erven and Cugliari, 2015, and
Wickramasuriya et al., 2019), or the equivalent structural approach by Hyndman
et al. (2011). Moreover, the classic bottom-up approach is available.
}
\details{
In case of non-negativity constraints, there are two ways:
\enumerate{
  \item \code{sol = "direct"} and \code{nn = TRUE}: the base forecasts
  will be reconciled at first without non-negativity constraints, then, if negative reconciled
  values are present, the \code{"osqp"} solver is used.
  \item \code{sol = "osqp"} and \code{nn = TRUE}: the base forecasts will be
  reconciled through the \code{"osqp"} solver.
}
}
\examples{
data(FoReco_data)
# monthly base forecasts
id <- which(simplify2array(strsplit(colnames(FoReco_data$base),
                                    split = "_"))[1, ] == "k1")
mbase <- t(FoReco_data$base[, id])
# monthly residuals
id <- which(simplify2array(strsplit(colnames(FoReco_data$res),
                                    split = "_"))[1, ] == "k1")
mres <- t(FoReco_data$res[, id])
obj <- htsrec(mbase, C = FoReco_data$C, comb = "shr", res = mres)

}
\references{
Byron, R.P. (1978), The estimation of large social accounts matrices,
\emph{Journal of the Royal Statistical Society A}, 141, 3, 359-367.

Di Fonzo, T., Girolimetto, D. (2020), Cross-Temporal Forecast Reconciliation:
Optimal Combination Method and Heuristic Alternatives, Department of Statistical
Sciences, University of Padua, \href{https://arxiv.org/abs/2006.08570}{arXiv:2006.08570}.

Hyndman, R.J., Ahmed, R.A., Athanasopoulos, G., Shang, H.L.(2011),
Optimal combination forecasts for hierarchical time series,
\emph{Computational Statistics & Data Analysis}, 55, 9, 2579-2589.

\enc{Sch채fer}{Schafer}, J.L., Opgen-Rhein, R., Zuber, V., Ahdesmaki, M.,
Duarte Silva, A.P., Strimmer, K. (2017), \emph{Package `corpcor'}, R
package version 1.6.9 (April 1, 2017), \href{https://CRAN.R-project.org/package=corpcor}{https://CRAN.R-project.org/package= corpcor}.

\enc{Sch채fer}{Schafer}, J.L., Strimmer, K. (2005), A Shrinkage Approach to Large-Scale Covariance
Matrix Estimation and Implications for Functional Genomics, \emph{Statistical
Applications in Genetics and Molecular Biology}, 4, 1.

Stellato, B., Banjac, G., Goulart, P., Bemporad, A., Boyd, S. (2018). OSQP:
An Operator Splitting Solver for Quadratic Programs, \href{https://arxiv.org/abs/1711.08013}{arXiv:1711.08013}.

Stellato, B., Banjac, G., Goulart, P., Boyd, S., Anderson, E. (2019), OSQP:
Quadratic Programming Solver using the 'OSQP' Library, R package version 0.6.0.3
(October 10, 2019), \href{https://CRAN.R-project.org/package=osqp}{https://CRAN.R-project.org/package=osqp}.

van Erven, T., Cugliari, J. (2015), Game-theoretically Optimal Reconciliation
of Contemporaneous Hierarchical Time Series Forecasts, in Antoniadis, A.,
Poggi, J.M., Brossat, X. (eds.), \emph{Modeling and Stochastic Learning for
Forecasting in High Dimensions}, Berlin, Springer, 297-317.

Wickramasuriya, S.L., Athanasopoulos, G., Hyndman, R.J. (2019), Optimal forecast
reconciliation for hierarchical and grouped time series through trace minimization,
\emph{Journal of the American Statistical Association}, 114, 526, 804-819.
}
\keyword{reconciliation}
