% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreco_info.R
\name{new_foreco_info}
\alias{new_foreco_info}
\alias{print.foreco_info}
\title{Low-level construction for reconcilied forecasts attribute foreco_info class}
\usage{
new_foreco_info(x = list())
}
\arguments{
\item{x}{A list of information related to the reconcilied forecasts.}
}
\description{
\code{new_foreca_info()} is the contructor for the \code{foreca_info} class, which acompany
the output from the reconciliation functions in the attribute \code{FoReco}.
This is exported for extension purposes and for expert use only.
}
\examples{
new_foreco_info(list(
  framework = "Cross-sectional",
  forecast_horizon = 1,
  comb = "shr",
  cs_n = 3,
  rfun = "csrec"
))
}
