% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reco_gauss.R
\name{temvn}
\alias{temvn}
\title{Temporal Gaussian probabilistic reconciliation}
\usage{
temvn(base, agg_order, tew = "sum", comb = "ols", res = NULL,
      approach = "proj", comb_base = comb, reduce_form = FALSE, ...)
}
\arguments{
\item{base}{A (\eqn{h(k^\ast + m) \times 1}) numeric vector containing the
base forecasts to be reconciled, ordered from lowest to highest frequency;
\eqn{m} is the maximum aggregation order, \eqn{k^\ast} is the sum of a
chosen subset of the \eqn{p - 1} factors of \eqn{m} (excluding \eqn{m}
itself) and \eqn{h} is the forecast horizon for the lowest frequency
time series.}

\item{agg_order}{Highest available sampling frequency per seasonal cycle
(max. order of temporal aggregation, \eqn{m}), or a vector representing a
subset of \eqn{p} factors of \eqn{m}.}

\item{tew}{A string specifying the type of temporal aggregation. Options
include: "\code{sum}" (simple summation, \emph{default}), "\code{avg}"
(average), "\code{first}" (first value of the period), and "\code{last}"
(last value of the period).}

\item{comb}{A string specifying the reconciliation method. For a complete
list, see \link{tecov}.}

\item{res}{A (\eqn{N(k^\ast+m) \times 1}) optional numeric vector containing
the in-sample residuals or validation errors ordered from the lowest
frequency to the highest frequency. This vector is used to compute come
covariance matrices.}

\item{approach}{A string specifying the approach used to compute the
reconciled forecasts. Options include:
\itemize{
\item "\code{proj}" (\emph{default}): Projection approach according
to Byron (1978, 1979).
\item "\code{strc}": Structural approach as proposed by Hyndman
et al. (2011).
\item "\code{proj_osqp}": Numerical solution using
\href{https://osqp.org/}{\pkg{osqp}} for projection approach.
\item "\code{strc_osqp}": Numerical solution using
\href{https://osqp.org/}{\pkg{osqp}} for structural approach.
}}

\item{comb_base}{A string specifying the reconciliation method. For a
complete list, see \link{tecov}.}

\item{reduce_form}{A logical parameter indicating whether the function
should return the full distribution (\code{FALSE}, \emph{default}) or
only the distribution corresponding to the high-frequency time series
(\code{TRUE}).}

\item{...}{
  Arguments passed on to \code{\link[=tecov]{tecov}}
  \describe{
    \item{\code{mse}}{If \code{TRUE} (\emph{default}) the errors used to compute
the covariance matrix are not mean-corrected.}
    \item{\code{shrink_fun}}{Shrinkage function of the covariance matrix,
\link{shrink_estim} (\emph{default})}
  }}
}
\value{
A \link[distributional:dist_multivariate_normal]{distributional::dist_multivariate_normal} object.
}
\description{
This function performs temporal probabilistic forecast reconciliation
assuming a multivariate normal base forecast distribution (Girolimetto et
al., 2024) for a single time series using temporal hierarchies
(Athanasopoulos et al., 2017).
}
\examples{
set.seed(123)
# (7 x 1) base forecasts vector (simulated), m = 4
base <- rnorm(7*2, rep(c(20, 10, 5), 2*c(1, 2, 4)))
# (70 x 1) in-sample residuals vector (simulated)
res <- rnorm(70)

m <- 4 # from quarterly to annual temporal aggregation
reco_dist <- terec(base = base, agg_order = m, comb = "wlsv", res = res)

}
\references{
Athanasopoulos, G., Hyndman, R.J., Kourentzes, N. and Petropoulos, F. (2017),
Forecasting with Temporal Hierarchies, \emph{European Journal of Operational
Research}, 262, 1, 60-74. \doi{10.1016/j.ejor.2017.02.046}

Byron, R.P. (1978), The estimation of large social account matrices,
\emph{Journal of the Royal Statistical Society, Series A}, 141, 3, 359-367.
\doi{10.2307/2344807}

Byron, R.P. (1979), Corrigenda: The estimation of large social account
matrices, \emph{Journal of the Royal Statistical Society, Series A}, 142(3),
405. \doi{10.2307/2982515}

Girolimetto, D., Athanasopoulos, G., Di Fonzo, T. and Hyndman, R.J. (2024),
Cross-temporal probabilistic forecast reconciliation: Methodological and
practical issues. \emph{International Journal of Forecasting}, 40, 3,
1134-1151. \doi{10.1016/j.ijforecast.2023.10.003}

Hyndman, R.J., Ahmed, R.A., Athanasopoulos, G. and Shang, H.L. (2011),
Optimal combination forecasts for hierarchical time series,
\emph{Computational Statistics & Data Analysis}, 55, 9, 2579-2589.
\doi{10.1016/j.csda.2011.03.006}
}
\seealso{
Probabilistic reconciliation:
\code{\link{csmvn}()},
\code{\link{cssmp}()},
\code{\link{ctmvn}()},
\code{\link{ctsmp}()},
\code{\link{tesmp}()}

Temporal framework:
\code{\link{teboot}()},
\code{\link{tebu}()},
\code{\link{tecov}()},
\code{\link{telcc}()},
\code{\link{temo}()},
\code{\link{terec}()},
\code{\link{tesmp}()},
\code{\link{tetd}()},
\code{\link{tetools}()}
}
\concept{Framework: temporal}
\concept{Reco: probabilistic}
