% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reco_sample.R
\name{tesmp}
\alias{tesmp}
\title{Temporal probabilistic reconciliation (sample approach)}
\usage{
tesmp(sample, agg_order, fun = terec, ...)
}
\arguments{
\item{sample}{A (\eqn{L \times h(k^\ast + m)}) numeric matrix containing the base forecasts
samples to be reconciled; \eqn{m} is the max aggregation order, \eqn{k^\ast} is the sum of
(a subset of) (\eqn{p-1}) factors of \eqn{m}, excluding \eqn{m}, \eqn{h} is the forecast
horizon for the lowest frequency time series, and \eqn{L} is the sample size.}

\item{agg_order}{Highest available sampling frequency per seasonal cycle
(max. order of temporal aggregation, \eqn{m}), or a vector representing a
subset of \eqn{p} factors of \eqn{m}.}

\item{fun}{A string specifying the reconciliation function to be used,
as implemented in \pkg{FoReco}.}

\item{...}{Arguments passed on to \code{fun}}
}
\value{
A \link[distributional:dist_sample]{distributional::dist_sample} object.
}
\description{
Temporal probabilistic reconciliation (sample approach)
}
\examples{
set.seed(123)
m <- 4 # from quarterly to annual temporal aggregation

# (100 x 14) base forecasts sample matrix (simulated), m = 4, h = 2
sample <- t(sapply(1:100, function(x) rnorm(14, rep(c(20, 10, 5), 2*c(1, 2, 4)))))
# (70 x 1) in-sample residuals vector (simulated)
res <- rnorm(70)

# Optimal cross-sectional probabilistic reconciliation
reco_dist_opt <- tesmp(sample, agg_order = m, res = res, comb = "shr")

# Bottom-up probabilistic reconciliation
reco_dist_bu <- tesmp(sample[,-c(1:6)], agg_order = m, fun = tebu)

# Level conditional coherent probabilistic reconciliation
reco_dist_lcc <- tesmp(sample, agg_order = m, fun = telcc)

}
\references{
Girolimetto, D., Athanasopoulos, G., Di Fonzo, T. and Hyndman, R.J. (2024),
Cross-temporal probabilistic forecast reconciliation: Methodological and
practical issues. \emph{International Journal of Forecasting}, 40, 3, 1134-1151.
\doi{10.1016/j.ijforecast.2023.10.003}

Panagiotelis, A., Gamakumara, P., Athanasopoulos, G. and Hyndman, R.J. (2023),
Probabilistic forecast reconciliation: Properties, evaluation and score optimisation,
\emph{European Journal of Operational Research} 306(2), 693–706.
\doi{http://dx.doi.org/10.1016/j.ejor.2022.07.040}
}
\seealso{
Probabilistic reconciliation:
\code{\link{csmvn}()},
\code{\link{cssmp}()},
\code{\link{ctmvn}()},
\code{\link{ctsmp}()},
\code{\link{temvn}()}

Temporal framework:
\code{\link{teboot}()},
\code{\link{tebu}()},
\code{\link{tecov}()},
\code{\link{telcc}()},
\code{\link{temo}()},
\code{\link{temvn}()},
\code{\link{terec}()},
\code{\link{tetd}()},
\code{\link{tetools}()}
}
\concept{Framework: temporal}
\concept{Reco: probabilistic}
