% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{ctboot}
\alias{ctboot}
\title{Cross-temporal joint block bootstrap}
\usage{
ctboot(model_list, boot_size, agg_order, block_size = 1, seed = NULL,
       xreg = NULL, ...)
}
\arguments{
\item{model_list}{A list of \eqn{p} elements, one for each temporal level
ordered from the lowest frequency (most temporally aggregated) to
the highest frequency. Each elements is a list with the \eqn{n} base
forecasts models for each cross-sectional series. A \code{simulate()}
function for each model has to be available and implemented according
to the package
\href{https://CRAN.R-project.org/package=forecast}{\pkg{forecast}},
with the following mandatory parameters: \emph{object}, \emph{innov},
\emph{future}, and \emph{nsim}.}

\item{boot_size}{The number of bootstrap replicates.}

\item{agg_order}{Highest available sampling frequency per seasonal cycle
(max. order of temporal aggregation, \eqn{m}), or a vector representing a
subset of \eqn{p} factors of \eqn{m}.}

\item{block_size}{Block size of the bootstrap, which is typically equivalent
to the forecast horizon for the most temporally aggregated series.}

\item{seed}{An integer seed.}

\item{xreg}{An optional 3-d numeric array of dimensions
(\eqn{n \times \text{boot\_size}(k^\ast+m) \times N_{xreg}})
containing the new values of \code{xreg} to be used for forecasting
ordered from the lowest frequency to the highest frequency (columns) for
each variable (rows). It can contain \code{NA}s.}

\item{...}{Additional arguments for the \code{simulate()} function.}
}
\value{
A list with two elements: the seed used to sample the errors and a
list with \eqn{\text{boot\_size}} matrix of size
(\eqn{n\times(k^\ast+m)\text{block\_size}}) matrix.
}
\description{
Joint block bootstrap for generating probabilistic base forecasts that take
into account the correlation between variables at different temporal
aggregation orders (Girolimetto et al. 2023).
}
\references{
Girolimetto, D., Athanasopoulos, G., Di Fonzo, T. and Hyndman, R.J. (2024),
Cross-temporal probabilistic forecast reconciliation: Methodological and
practical issues. \emph{International Journal of Forecasting}, 40, 3,
1134-1151. \doi{10.1016/j.ijforecast.2023.10.003}
}
\seealso{
Bootstrap samples:
\code{\link{csboot}()},
\code{\link{teboot}()}

Cross-temporal framework:
\code{\link{ctbu}()},
\code{\link{ctcov}()},
\code{\link{ctlcc}()},
\code{\link{ctmo}()},
\code{\link{ctmvn}()},
\code{\link{ctrec}()},
\code{\link{ctsmp}()},
\code{\link{cttd}()},
\code{\link{cttools}()},
\code{\link{iterec}()},
\code{\link{tcsrec}()}
}
\concept{Bootstrap samples}
\concept{Framework: cross-temporal}
