% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoReco-package.R
\docType{package}
\name{FoReco-package}
\alias{FoReco}
\alias{FoReco-package}
\title{FoReco: Forecast Reconciliation}
\description{
Classical (bottom-up and top-down), optimal combination and heuristic point (Di Fonzo and Girolimetto, 2023 \doi{10.1016/j.ijforecast.2021.08.004}) and probabilistic (Girolimetto et al. 2023 \doi{10.1016/j.ijforecast.2023.10.003}) forecast reconciliation procedures for linearly constrained time series (e.g., hierarchical or grouped time series) in cross-sectional, temporal, or cross-temporal frameworks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/daniGiro/FoReco}
  \item \url{https://danigiro.github.io/FoReco/}
  \item Report bugs at \url{https://github.com/daniGiro/FoReco/issues}
}

}
\author{
\strong{Maintainer}: Daniele Girolimetto \email{daniele.girolimetto@unipd.it} (\href{https://orcid.org/0000-0001-9387-1232}{ORCID}) [funder]

Authors:
\itemize{
  \item Tommaso Di Fonzo (\href{https://orcid.org/0000-0003-3388-7827}{ORCID}) [funder]
}

}
\keyword{package}
