% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoReco2ts.R
\name{FoReco2ts}
\alias{FoReco2ts}
\title{Reconciled forecasts matrix/vector to time-series class}
\usage{
FoReco2ts(recf, ...)
}
\arguments{
\item{recf}{(\mjseqn{h(k^\ast + m) \times 1}) reconciled forecasts vector from \link[FoReco]{thfrec},
(\mjseqn{h \times n}) reconciled forecasts matrix from \link[FoReco]{htsrec} or
(\mjseqn{n \times h(k^\ast + m)}) reconciled forecasts matrix from \link[FoReco]{octrec},
\link[FoReco]{tcsrec}, \link[FoReco]{cstrec}, \link[FoReco]{iterec},
\link[FoReco]{ctbu}.}

\item{...}{optional arguments to \link[stats]{ts} (i.e. starting date);
frequency is required only for the cross-sectional case.}
}
\value{
A list of class \code{"ts"} objects
}
\description{
\loadmathjax
Function to transform the matrix/vector of reconciled forecasts
(\code{recf} from \link[FoReco]{htsrec}, \link[FoReco]{thfrec}, \link[FoReco]{tdrec},
\link[FoReco]{octrec}, \link[FoReco]{lccrec}, \link[FoReco]{tcsrec}, \link[FoReco]{cstrec},
\link[FoReco]{iterec}, \link[FoReco]{ctbu}) into a list of time series objects.
}
\examples{
data(FoReco_data)
# Cross-temporal framework
oct_recf <- octrec(FoReco_data$base, m = 12, C = FoReco_data$C,
                   comb = "bdshr", res = FoReco_data$res)$recf
obj_oct <- FoReco2ts(recf = oct_recf, start = c(15, 1))

# Cross-sectional framework
# monthly base forecasts
id <- which(simplify2array(strsplit(colnames(FoReco_data$base),
                                    split = "_"))[1, ] == "k1")
mbase <- t(FoReco_data$base[, id])
# monthly residuals
id <- which(simplify2array(strsplit(colnames(FoReco_data$res),
                                    split = "_"))[1, ] == "k1")
mres <- t(FoReco_data$res[, id])
hts_recf <- htsrec(mbase, C = FoReco_data$C, comb = "shr", res = mres)$recf
obj_hts <- FoReco2ts(recf = hts_recf, start = c(15, 1), frequency = 12)

# Temporal framework
# top ts base forecasts ([lowest_freq' ...  highest_freq']')
topbase <- FoReco_data$base[1, ]
# top ts residuals ([lowest_freq' ...  highest_freq']')
topres <- FoReco_data$res[1, ]
thf_recf <- thfrec(topbase, m = 12, comb = "acov", res = topres)$recf
obj_thf <- FoReco2ts(recf = thf_recf, start = c(15, 1))

}
\seealso{
Other utilities: 
\code{\link{Cmatrix}()},
\code{\link{commat}()},
\code{\link{ctf_tools}()},
\code{\link{hts_tools}()},
\code{\link{oct_bounds}()},
\code{\link{score_index}()},
\code{\link{shrink_estim}()},
\code{\link{srref}()},
\code{\link{thf_tools}()},
\code{\link{ut2c}()}
}
\concept{utilities}
\keyword{utilities}
